Pathfinding Fix
I made a custom grid and a custom quad tree. Each cell uses a collider to check for collision when it is being generated, if anything collides and its not with a certain tag this cell will be blocked, the grid size can be customized and the cell size can also be customized.