Fixed enemy Spawn and improved Pathfinding
Custim path finding with a custom quad tree and custom node grid. Enemies can spawn at different points and move towards points and players, they dont collider with sprites but they avoid them. If a sprite is the target the target will change to next to the sprite otherwise they wont be able to move to the target. When they reach a target they switch to attack state but at the moment this does nothing. My ram on my compter is broken because I have bad luck or something but I think it works performance wise on other computers.