AI
The pathfinding only takes time during start, when you reset the night to day it doesnt take long.
the enemies that has a dynamic target stops every time they caluclate a new path, they do this every 3 seconds atm. this cant be changed. But I wll implement flock some time soon and then maybe that can be fixed. collision on enemies is impossible cus when they are moved they need to find a new path and if the path is long they would need many searches which destroys performance. this might also be fixed with flock.